# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CategoryResponse:
    properties:
      coffee_shop_id:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
...
  dto.CreateCategory:
    properties:
      description:
        type: string
      title:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - title
    type: object
...
  dto.CreateIdeaRequest:
    properties:
      category_id:
        type: string
      coffee_shop_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      title:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.IdeaResponse:
    properties:
      category_id:
        type: string
      coffee_shop_id:
        type: string
      creator_id:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      likes:
        type: integer
      status_id:
        type: string
      title:
        type: string
    type: object
...
  dto.UpdateCategory:
    properties:
      description:
        type: string
      title:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - title
    type: object
...
  dto.UpdateIdeaRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      image_url:
        type: string
      status_id:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /coffee-shops/{id}/categories:
    get:
      description: Get a list of all categories for a given coffee shop with optional
        pagination
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get categories by coffee shop
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category for a coffee shop
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Category information
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new category
      tags:
      - categories
  /coffee-shops/{id}/categories/{category_id}:
    delete:
      description: Delete a category by ID
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a category
      tags:
      - categories
    get:
      description: Get category details by ID
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update category details for the given ID
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      - description: Category update information
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a category
      tags:
      - categories
  /coffee-shops/{id}/ideas:
    get:
      description: Get a list of all ideas for a given coffee shop with optional pagination
      parameters:
      - description: Coffee Shop ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Sort order
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.IdeaResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all ideas by shop
      tags:
      - ideas
...
  /ideas:
    post:
      consumes:
      - application/json
      description: Create a new idea for a coffee shop
      parameters:
      - description: Idea information
        in: body
        name: idea
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIdeaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.IdeaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new idea
      tags:
      - ideas
  /ideas/{id}:
    delete:
      description: Delete an idea by ID
      parameters:
      - description: Idea ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete idea by ID
      tags:
      - ideas
    get:
      description: Get idea details by ID
      parameters:
      - description: Idea ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IdeaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get idea by ID
      tags:
      - ideas
    put:
      consumes:
      - application/json
      description: Update idea details for the given ID
      parameters:
      - description: Idea ID
        in: path
        name: id
        required: true
        type: string
      - description: Idea update information
        in: body
        name: idea
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateIdeaRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update idea by ID
      tags:
      - ideas
...
  /users/me/ideas:
    get:
      description: Get a list of all ideas for a given user with optional pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Sort order
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.IdeaResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all ideas by user
      tags:
      - ideas
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"